refactor(getFromFile): change facade method name to `get`
